+++ packages/breadcrumbs/Breadcrumbs.d.ts
@@ -6,6 +6,7 @@ export interface BreadcrumbsProps {
     crumbs?: Array<Crumb>;
     active: string;
     emptyState?: string;
+    homeUrl?: string;
 }
 
 declare const Breadcrumbs: React.StatelessComponent<BreadcrumbsProps>;