+++ packages/date/typings/DateRangeField.d.ts
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { DateRangeProps } from './DateRange';
 
 export interface DateRangeFieldProps extends DateRangeProps {
-  label?: string;
+  label?: React.ReactNode;
 }
 
 declare class DateRangeField extends React.Component<DateRangeFieldProps> {}