+++ packages/form/typings/RadioGroup.d.ts
@@ -3,8 +3,9 @@ import { FormGroupProps } from './FormGroup';
 
 interface RadioGroupProps extends FormGroupProps {
   name: string;
-  label?: Node;
+  label?: React.ReactNode;
   onChange?: (value: any) => void;
+  inline?: boolean | false;
 }
 
 declare class RadioGroup extends React.Component<RadioGroupProps> {}