+++ packages/form/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2](https://github.com/Availity/availity-react/compare/@availity/form@0.4.1...@availity/form@0.4.2) (2019-09-23)
+
+**Note:** Version bump only for package @availity/form
+
+
+
+
+
+## [0.4.1](https://github.com/Availity/availity-react/compare/@availity/form@0.4.0...@availity/form@0.4.1) (2019-09-19)
+
+**Note:** Version bump only for package @availity/form
+
+
+
+
+
+# [0.4.0](https://github.com/Availity/availity-react/compare/@availity/form@0.3.3...@availity/form@0.4.0) (2019-09-13)
+
+
+### Features
+
+* **form:** adds inline radio control to RadioGroup ([a9c344c](https://github.com/Availity/availity-react/commit/a9c344c))
+* **form:** update type definitions ([e6e8ce9](https://github.com/Availity/availity-react/commit/e6e8ce9))
+
+
+
+
+
 ## [0.3.3](https://github.com/Availity/availity-react/compare/@availity/form@0.3.2...@availity/form@0.3.3) (2019-09-06)