+++ packages/select/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.8](https://github.com/Availity/availity-react/compare/@availity/select@0.4.7...@availity/select@0.4.8) (2019-09-23)
+
+**Note:** Version bump only for package @availity/select
+
+
+
+
+
+## [0.4.7](https://github.com/Availity/availity-react/compare/@availity/select@0.4.6...@availity/select@0.4.7) (2019-09-19)
+
+
+### Bug Fixes
+
+* **select:** fixes multiple API calls on component mount ([fd45b4a](https://github.com/Availity/availity-react/commit/fd45b4a))
+
+
+
+
+
+## [0.4.6](https://github.com/Availity/availity-react/compare/@availity/select@0.4.5...@availity/select@0.4.6) (2019-09-19)
+
+**Note:** Version bump only for package @availity/select
+
+
+
+
+
+## [0.4.5](https://github.com/Availity/availity-react/compare/@availity/select@0.4.4...@availity/select@0.4.5) (2019-09-13)
+
+**Note:** Version bump only for package @availity/select
+
+
+
+
+
+## [0.4.4](https://github.com/Availity/availity-react/compare/@availity/select@0.4.3...@availity/select@0.4.4) (2019-09-13)
+
+
+### Bug Fixes
+
+* **select:** use valueKey when creating new options ([6d5c113](https://github.com/Availity/availity-react/commit/6d5c113))
+
+
+
+
+
 ## [0.4.3](https://github.com/Availity/availity-react/compare/@availity/select@0.4.2...@availity/select@0.4.3) (2019-09-06)
 
 **Note:** Version bump only for package @availity/select